Update(scripts): more scripts

+++ tta/tafas.py
@@ -336,7 +336,7 @@ class Adapter(nn.Module):
         self.model.eval()
     
     def _calculate_period_and_batch_size(self, enc_window_first):
-        fft_result = torch.fft.rfft(enc_window_first - enc_window_first.mean(dim=0), dim=0).to(enc_window_first.dtype)
+        fft_result = torch.fft.rfft(enc_window_first - enc_window_first.mean(dim=0), dim=0)
         # amplitude = torch.abs(fft_result)
         amplitude = torch.sqrt(fft_result.real.pow(2) + fft_result.imag.pow(2))
         power = torch.mean(amplitude ** 2, dim=0)
